Removing WebFinger support
It's hard to implement any kind of WebFinger support because there is no clear accepted format for it. The original format that's the most widely used doesn't conform to the final format of "Web Host Metadata" and when people finally were about to update to use the final format and especially to implement support for the JRD format then WebFinger suddenly appeared in the IETF discussions as a specification of its own which decided that WebFinger should not at all just be a simple extension of the "Web Host Metadata" specification but rather a specification of its own with its own place in the .well-known directory that doesn't use XRD but (perhaps?) just JRD and which also changes(?) some further things like how HTTPS should be used and if you can serve WebFinger from a static server or not(?).

All in all: What was an easy fun thing to add to a site as a small step towards a more open web has turned into something very confusing that there is no real way of supporting at all right now as far as I can see - especially not on GitHub Pages where I'm now moving this site. This is sad because WebFinger was (is?) an important part of the Federated Social Web movement which is a movement which success I think gets more important for every day and which I think can't afford nit-picky specification work in favor of real world deployment of the specs.